+++ client/src/components/index.js
@@ -1,5 +1,4 @@
 // EMPLOYER
-import EmployerLogin from './employer/EmployerLogin';
 import EmployerProfile from './employer/EmployerProfile';
 import EmployerRegister from './employer/EmployerRegister';
 import EmployerMatches from './employer/EmployerMatches';
@@ -8,16 +7,15 @@ import EmployerUploadJob from './employer/EmployerUploadJob';
 import EmployerBrowseSeekers from './employer/EmployerBrowseSeekers';
 // SEEKER
 import SeekerRegister from './seeker/SeekerRegister';
-import SeekerLogin from './seeker/SeekerLogin';
 import SeekerProfile from './seeker/SeekerProfile';
 import SeekerMatches from './seeker/SeekerMatches';
 import SeekerMatchView from './seeker/SeekerMatchView';
 import SeekerBrowseJobs from './seeker/SeekerBrowseJobs';
 // LANDING
-import LandingStart from './landing/LandingStart';
-import LandingLogin from './landing/LandingLogin';
-import LandingRegister from './landing/LandingRegister';
-import LoginOrRegister from './landing/LoginOrRegister';
+import LandingMain from './landing/LandingMain';
+import Toggle from './landing/Toggle';
+import Login from './landing/LandingLogin';
+
 // MESSENGER
 import MessageHistory from './messenger/MessageHistory';
 import MessageSideBar from './messenger/MessageSideBar';
@@ -27,7 +25,6 @@ import MessageCompose from './messenger/MessageCompose';
 export {
   // EMPLOYER
   EmployerRegister,
-  EmployerLogin,
   EmployerProfile,
   EmployerMatches,
   EmployerMatchView,
@@ -35,7 +32,6 @@ export {
   EmployerBrowseSeekers,
   // SEEKER
   SeekerRegister,
-  SeekerLogin,
   SeekerProfile,
   SeekerMatches,
   SeekerMatchView,
@@ -45,8 +41,7 @@ export {
   MessageSideBar,
   MessageCompose,
   // OTHER
-  LandingStart,
-  LandingLogin,
-  LandingRegister,
-  LoginOrRegister,
+  LandingMain,
+  Toggle,
+  Login
 };
