+++ client/src/components/seeker/SeekerRegister.js
@@ -1,179 +1,75 @@
-import React, { Component } from 'react';
-import { withRouter } from 'react-router-dom';
-import { connect } from 'react-redux';
-
-import { registerUser } from '../../actions';
-
+import React from 'react';
 import {
-  RegisterContainer,
-  Form,
-  InputContainer,
-  InputBox,
-  InputTextarea,
-  Notification,
-  ButtonsContainer,
-  Button,
+  LandingForm,
+  LandingFormLabel,
+  LandingInput50,
+  LandingInput100,
+  LandingTextarea,
+  LandingSectionUser,
+  LandingSectionInfo,
+  LandingSectionCredentials,
+  LandingSubmitButton
 } from '../styles';
 
-class SeekerRegister extends Component {
-  state = {
-    firstName: '',
-    lastName: '',
-    email: '',
-    desiredTitle: '',
-    summary: '',
-    topSkills: [],
-    password: '',
-    passwordLengthOk: true,
-    passwordMatch: true,
-    summaryLengthOK: true,
-  };
-
-  handleChange({ target }) {
-    const { name, value } = target;
-    let { summaryLengthOK, password, confirmPassword } = this.state;
-    let splitSkills;
-    // splitting skills array by commas and ignoring whitespace at substring edges
-    if (name === "topSkills") {
-      splitSkills = value.split(/, */);
-    }
-    // live feedback for password match and summary length
-    switch (name) {
-      case 'password':
-        password = value;
-        break;
-      case 'confirmPassword':
-        confirmPassword = value;
-        break;
-      case 'summary':
-        summaryLengthOK = value.length <= 256;
-        break;
-      default:
-        break;
-    }
-    // check password length and match
-    const passwordLengthOk = !password || password.length >= 8;
-    const passwordMatch = password === confirmPassword;
-    this.setState({
-      summaryLengthOK,
-      passwordLengthOk,
-      passwordMatch,
-      [name]: splitSkills || value,
-    });
-  }
-
-
-  submitHandler(event) {
-    event.preventDefault();
-    const {
-      firstName,
-      lastName,
-      desiredTitle,
-      summary,
-      topSkills,
-      experience,
-      education,
-      password,
-      email,
-      passwordLengthOk,
-      passwordMatch,
-    } = { ...this.state };
-
-    if (!passwordLengthOk) {
-    } else if (!passwordMatch) {
-    } else if (!desiredTitle || !summary || !topSkills
-      || !firstName || !lastName || !experience || !education || !password || !email) {
-    } else {
-      this.props.registerUser(this.state, 'jobseeker');
-    }
-  }
-
-  render() {
-    return (
-    <RegisterContainer>
-      <Form onSubmit={this.submitHandler.bind(this)}>
-        <InputContainer row signup separate>
-          <InputBox 
-            type="text" 
-            name="firstName" 
-            placeholder="Your first name" 
-            onChange={this.handleChange.bind(this)} 
-          />
-          <InputBox 
-            type="text" 
-            name="lastName" 
-            placeholder="Your last name" 
-            onChange={this.handleChange.bind(this)} 
-          />
-        </InputContainer>
-        <InputContainer row signup>
-          <InputBox 
-            type="email" 
-            name="email" 
-            placeholder="Your Email address" 
-            onChange={this.handleChange.bind(this)} 
-          />
-          <InputBox 
-            type="text" 
-            name="desiredTitle" 
-            placeholder="What is your desired title?" 
-            onChange={this.handleChange.bind(this)} 
-          />
-        </InputContainer>
-        <InputContainer>
-          <InputTextarea large 
-            type="text"
-            name="summary" 
-            placeholder="Summarize yourself" 
-            onChange={this.handleChange.bind(this)} 
-           />
-          <Notification alert={!this.state.summaryLengthOK}>
-            {this.state.summary.length}
-          </Notification>
-          <InputBox 
-            type="text" 
-            name="topSkills" 
-            placeholder="Select your top skills, max five" onChange={this.handleChange.bind(this)} 
-          />
-          <InputTextarea large 
-            type="text" 
-            name="experience" 
-            placeholder="List your experience (Job Title, YearStarted - YearEnded/Current)" 
-            onChange={this.handleChange.bind(this)} 
-          />       
-          <InputBox 
-            type="text" 
-            name="education" 
-            placeholder="Educational Experience (School, Year Graduated)" 
-            onChange={this.handleChange.bind(this)} 
-          />
-        </InputContainer>
-        <InputContainer row signup>
-          <InputBox 
-            type="password" 
-            name="password" 
-            placeholder="Password" 
-            onChange={this.handleChange.bind(this)} 
-          />
-          <InputBox 
+const SeekerRegister = (props) => {
+  return (
+    <LandingForm onSubmit={(e) => props.register(e)}>
+      <LandingFormLabel> SEEKER SIGNUP</LandingFormLabel> 
+      <LandingSectionUser>
+        <LandingInput50 
+          name="firstName" 
+          placeholder="Your First Name"
+        />
+        <LandingInput50 
+          name="lastName"
+          placeholder="Your Last Name"
+        />
+        <LandingInput50 
+          name="email"
+          type="email"
+          placeholder="Your Email Address"
+        />
+        <LandingInput50 
+          name="desiredTitle" 
+          placeholder="Your Desired Title"
+        />
+        <LandingTextarea 
+          name="summary" 
+          placeholder="Summarize Yourself" 
+        />
+      </LandingSectionUser>
+      <LandingSectionInfo>
+        <LandingInput100 
+          name="topSkills" 
+          placeholder="Select You Top Skills (Max 5)"
+        />
+        <LandingTextarea 
+          name="experience"
+          placeholder="List Your Experience (Job Title, Year Started - Year Ended, Current"
+        />
+        <LandingInput100 
+          name="education"
+          placeholder="Educational Experience (School, Year Graduated)"
+        />
+      </LandingSectionInfo>
+      <LandingSectionCredentials>
+        <LandingInput50 
+          name="password" 
           type="password" 
-          name="confirmPassword" 
-          placeholder="Confirm Password" 
-          onChange={this.handleChange.bind(this)} 
-          />
-        </InputContainer>
-        <Notification alert>
-          {this.state.passwordLengthOk ? '' : 'Password is too short.'}
-        </Notification>
-        <Notification alert>
-          {this.state.passwordMatch ? '' : 'Passwords do not match.'}
-        </Notification>
-        <ButtonsContainer>
-          <Button type="submit">Find a Job</Button>
-        </ButtonsContainer>
-      </Form>
-      </RegisterContainer>);
-  }
+          placeholder="Password"
+        />
+        <LandingInput50 
+          name="confirm" 
+          type="password" 
+          placeholder="Confirm"
+        />
+      </LandingSectionCredentials>
+      <LandingSubmitButton 
+        type="submit" 
+        value="FIND A JOB"
+      />       
+    </LandingForm>
+  )
 }
 
-export default withRouter(connect(null, { registerUser })(SeekerRegister));
+export default SeekerRegister;
