+++ client/src/components/styles/index.js
@@ -296,7 +296,7 @@ export const DropDownArrow = styled.p`
 `;
 
 // ***************
-// T7 LANDING STYLES
+// LANDING STYLES
 // ***************
 export const Main = styled.main`
   width: 100%;
@@ -459,10 +459,6 @@ export const LandingToggleButton = styled.button`
 export const LandingLogo = styled.img`
   width: 80%;
 `;
-// ***************
-// LANDING STYLES
-// ***************
-
 export const BackgroundContainer = styled.div`
   display: flex;
   justify-content: center;
@@ -491,58 +487,7 @@ export const BackgroundContainer = styled.div`
     background-size: cover;
   }
 `;
-export const AppTitle = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center; 
-  width: 40%;
-  font-family: 'Roboto Black';
-  color: white;
-  text-shadow: 1px 1px 2px black;
-  font-size: 13vw;
-`;
-export const RegisterAndLoginContainer = styled.div`
-  display: flex;
-  width: 60%;
-  flex-direction: column; 
-  justify-content: flex-start;
-  padding: 5% 1%;
-  padding-bottom: 0;
-  background-color: rgba(171, 178, 225, .3);
-`;
-export const LoginContainer = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-export const RegisterContainer = styled.div`
-  width: 100%;
-  display: flex;
-  height: 90%;
-  flex-direction: column;
-  align-items: center;
-  overflow-y: scroll;
-  -webkit-overflow-scrolling: touch;
-`;
-export const SeekerOrEmployer = styled.div`
-  float: ${props => (props.right ? 'right' : 'left')};
-`;
-export const LandingButton = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 50%;
-  box-sizing: border-box;
-  height: ${props => (props.small ? '50px' : '75px')};
-  color: ${props => (props.selected ? 'white' : buttonColor)};
-  background-color: ${props => (props.selected ? buttonColor : 'white')};
-  border: ${props => (props.selected ? '1px solid black' : '6px inset lightgrey')};
-  margin: 1%;
-  font-size: ${props => (props.small ? '24px' : '32px')};
-  -webkit-transition-duration: 0.6s;
-  transition-duration: 0.6s;
-`;
+
 
 // LOADING
 export const Loading = styled.div`
