--- client/src/components/styles/landingStart.css
@@ -1,175 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Cardo');
-.LS-wrapper {
-  background: rgba(0,0,0,0.5);
-  opacity: 0.75;
-  width: 30%;
-  height: 100%;
-  min-height: 535px;
-  position: relative;
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  min-width: calc(300px + 2rem);
-  max-height: 100%;
-  overflow-y: auto;
-  transition: height 1s;
-}
-main > div > label {
-  
-}
-
-.LS-wrapper > label {
-  text-align: center;
-}
-
-.LS-wrapper > header {
-  border-bottom: 2px solid black; 
-  width: 100%;
-  display: flex;
-}
-.LS-wrapper > header > div {
-  background:rgb(2, 2, 108);
-  color: white;
-  width: 100%;
-  height: 100%;
-  padding: 1rem;
-  font-size: calc(0.9rem + 0.1vw);
-  font-weight: 700; 
-  text-align: center;
-  border-radius: 0.5rem 0.5rem 0 0;
-}
-
-.LS-login {
-  display: flex;
-  justify-content: center;  
-  flex-direction: column;
-  width: 100%;
-}
-
-.LS-login > input {
-  width: calc(100% - 0.5rem);
-  margin: 0.25rem;
-  padding: 0.5rem;
-  text-align: center;
-}
-
-.selected {
-  background: blue;
-}
-.unselected {
-  background: rgb(2, 2, 79);
-}
-
-.LS-wrapper > header > button {
-  border-top-left-radius: 0.5rem; 
-  border-top-right-radius: 0.5rem; 
-  border: 1px solid black;
-  border-bottom: none;
-  outline: none; 
-  padding: 0.2rem 0;
-  color: white; 
-  cursor: pointer;
-  font-size: calc(1rem + 0.75vw);
-  transition: background-color 0.75s;
-  flex: 1;
-}
-
-.LS-form {
-  background: whitesmoke;
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  border-radius: 0 0 0.5rem 0.5rem;
-  overflow-y: auto;  
-  overflow-x: hidden;  
-  font-size: calc(0.5rem + 0.75vw);
-}
-
-.LS-form > form{
-  width: 100%;
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  align-items : center;
-  overflow: auto;
-}
-
-
-
-.LS-form > span {
-  text-align: center;
-  padding-bottom: 1rem;
-}
-
-.LS-form > span > button {
-  border: none;
-  background: none;
-  color: blue;
-  outline: none;
-  cursor: pointer;
-}
-
-.LS-form > form > button, .submit {
-  width: calc(100% - 0.5rem);
-  margin: 0.25rem;
-  background: blue;
-  color: white;
-  border: none;
-  border-radius: 0.25rem;
-  cursor: pointer;
-  transition: background-color 0.45s;
-  padding: 1rem 0;
-
-}
-
-.LS-form > div > button:hover{
-  background: rgb(2, 2, 108);
-}
-
-.LS-user {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: center;
-}
-
-.LS-user > input {
-  width: calc(50% - 0.5rem);
-  margin: 0.25rem;
-  padding: 0.2rem;
-  word-wrap: wrap;
-}
-
-.LS-form textarea {
-  width: calc(100% - 0.5rem);
-  padding: 0.5rem;
-  min-height: 3rem;
-  resize: none;
-
-}
-
-.LS-info {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  align-items: center;
-}
-
-.LS-info > input {
-  width: calc(100% - 0.5rem);
-  margin: 0.25rem;
-  padding: 0.5rem;
-}
-
-.LS-credentials {
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: row;
-}
-
-.LS-credentials > input {
-  width: calc(50% - 0.5rem);
-  margin: 0.25rem;
-  padding: 0.5rem;
-}
